+++ src/components/account/accountShop/accountMyShopping.js
@@ -1,8 +1,9 @@
 import React, { useState } from "react";
-import DeliStateDropdown from "../common/deliStateDropdown";
-import PeriodDropdown from "../common/periodDropdown";
+import DeliStateDropdown from "../../common/dropdown/deliStateDropdown";
+import PeriodDropdown from "components/common/dropdown/periodDropdown";
+import MyReviewList from "../accountReview/myReviewList"; // request My review 말고 my product 같은걸로 바꿔서 props로 배송상태 주고 상태에 따라 띄우는거 어떤가
+
 import classes from "./accountMyShopping.module.css";
-import MyReviewList from "./myReviewList"; // request My review 말고 my product 같은걸로 바꿔서 props로 배송상태 주고 상태에 따라 띄우는거 어떤가
 
 function AccountMyShopping(props) {
   const [inqPeriod, setInqPeriod] = useState("");
@@ -12,7 +13,7 @@ function AccountMyShopping(props) {
 
   const DUMMY_DATA_UNREVIEWED = [
     {
-      id: 1,
+      id: 11,
       product_image:
         "http://imagescdn.gettyimagesbank.com/500/21/378/818/0/1326923672.jpg",
       product_brand: "시디즈",
@@ -20,7 +21,7 @@ function AccountMyShopping(props) {
       product_option: "블랙",
     },
     {
-      id: 1,
+      id: 12,
       product_image:
         "https://image.ohou.se/i/bucketplace-v2-development/uploads/productions/159719442642625646.jpg?gif=1&w=1280&h=1280&c=c",
       product_brand: "시디즈",
@@ -63,9 +64,10 @@ function AccountMyShopping(props) {
           <div className={classes.acctMyShopping_prdndelis}>
             {showInqPeriod && (
               <p className={classes.acctMyShopping_SelectedRslt}>
-                <p className={classes.acctMyShopping_selectedTextArea}>
-                  {inqPeriod}
-                </p>
+                {/* 에러나서 일단 없앰, 나중에 정식으로 손보자 */}
+                {/* <p className={classes.acctMyShopping_selectedTextArea}> */}
+                {inqPeriod}
+                {/* </p> */}
                 <button
                   className={classes.acctMyShopping_selectedRmvBtn}
                   onClick={clearPeriodValue}
@@ -76,9 +78,9 @@ function AccountMyShopping(props) {
             )}
             {showInqDeliState && (
               <p className={classes.acctMyShopping_SelectedRslt}>
-                <p className={classes.acctMyShopping_selectedTextArea}>
-                  {inqDeliState}
-                </p>
+                {/* <p className={classes.acctMyShopping_selectedTextArea}> */}
+                {inqDeliState}
+                {/* </p> */}
                 <button
                   className={classes.acctMyShopping_selectedRmvBtn}
                   onClick={clearDeliStateValue}